Moderator can accept/deny offers to be a speaker on a free topic

+++ src/main/java/com/conference/demo/dto/SpeakerOfferDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class SpeakerOfferDTO {
 
-    private int reportTopicId;
+    private long reportTopicId;
 }
